requirements.txt
requests~=2.32.5
selenium~=4.41.0
webdriver-manager
allure-pytest~=2.15.3
python-dotenv~=1.2.1
